requirements.txt
astroid==2.15.6
beautifulsoup4==4.9.3
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
dill==0.3.7
Flask==2.3.2
google-spell-checker==0.2.1
idna==3.4
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
lazy-object-proxy==1.9.0
Markdown==3.4.4
MarkupSafe==2.1.3
mccabe==0.7.0
md-to-html==0.7.3
mistune==0.8.4
platformdirs==3.10.0
pylint==2.17.5
PyPDF2==3.0.1
pyspellchecker==0.7.2
python-frontmatter==1.0.0
PyYAML==6.0.1
requests==2.31.0
soupsieve==2.4.1
tomlkit==0.12.1
urllib3==2.0.4
Werkzeug==2.3.6
wrapt==1.15.0
